requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
requests
sentry-sdk
mysql-connector-python
cryptography
drf-jwt
pyjwt
python-memcached
channels
daphne
django-filter
web3
